Move RkTagThrottleCollection class into its own files
